requirements.txt
requests
PyJWT
validate_email
sqlalchemy
flask
